Moved from verify to cluster package
