+++ pkg/e2e/e2e.go
@@ -56,7 +56,6 @@ import (
 	"github.com/openshift/osde2e/pkg/common/util"
 	"github.com/openshift/osde2e/pkg/debug"
 	"github.com/openshift/osde2e/pkg/e2e/routemonitors"
-	"github.com/openshift/osde2e/pkg/e2e/verify"
 	"github.com/openshift/osde2e/pkg/reporting/ginkgorep"
 )
 
@@ -161,7 +160,7 @@ func beforeSuite() bool {
 
 	//Creates a configmap for PrCheck to use as a psudo queue
 	if viper.GetBool(config.Tests.EnablePrCheck) {
-		if err := verify.PrCheckQueue(); err != nil {
+		if err := cluster.PrCheckQueue(); err != nil {
 			log.Printf("Failed to create queue for PrCheck: %v", err)
 			viper.Set(config.Tests.EnablePrCheck, false)
 			viper.Set(config.Cluster.HibernateAfterUse, false)
@@ -674,7 +673,7 @@ func cleanupAfterE2E(h *helper.H) (errors []error) {
 
 	// Cleanup PrCheckQueue after the cluster is back to default state
 	if viper.GetBool(config.Tests.EnablePrCheck) {
-		verify.PrCheckQueueCleanup()
+		cluster.PrCheckQueueCleanup()
 	}
 
 	// We need to clean up our helper tests manually.
